Write file to a temporary file to avoid file corruption when FS is full
When overwriting an existing file, it is truncated and it could then fail if the
filesystem is full. A valid file as thus been replaced by an empty one.
Using a temporary file for writing should prevent this behavior as writing to
the temporary file might fails before replacing the valid one.